refactor: extract exploration tool adapters
